// LandingPageSensora/lib/types/loja.ts
export type CategoriaResumo = {
  id: number;
  nome: string;
  slug: string;
};

export type Categoria = {
  id: number;
  nome: string;
  slug: string;
  descricao?: string | null;
};

export type CreateCategoriaPayload = {
  nome: string;
  descricao?: string;
};

export type UpdateCategoriaPayload = Partial<CreateCategoriaPayload>;

export type Produto = {
  id: number;
  nome: string;
  slug: string;
  descricao?: string | null;
  preco: number;
  quantidade: number;
  imagemUrl?: string | null;
  aroma?: string | null;
  ativo: boolean;
  destaque: boolean;
  categoriaId?: number | null;
  categoria?: CategoriaResumo | null;
};

export type CreateProdutoPayload = {
  nome: string;
  descricao?: string;
  aroma?: string;
  imagemUrl?: string;
  ativo?: boolean;
  destaque?: boolean;
  categoriaId?: number;
  preco: number;
  quantidade: number;
};